refactor: extract project health metrics calculation into ProjectMetricsCalculator
- move calculate_repository_metrics and the get_metrics calculation
  helpers out of ProjectTools into a new RedmineAiHelper::ProjectMetricsCalculator,
  fixing a Metrics/ClassLength RuboCop violation
- update ProjectAgent#append_repository_metrics to use the new calculator
  instead of calling ProjectTools directly
- split the corresponding unit tests into project_metrics_calculator_test.rb
  and update the project_agent_test.rb mock accordingly
- exclude the new file from Metrics/AbcSize and Metrics/CyclomaticComplexity,
  matching the existing exclusion for lib/redmine_ai_helper/tools/**/*
